fix Sials shit

+++ Backend/Models/User.cs
@@ -4,7 +4,7 @@ namespace Backend.Models
         public required int user_id { get; set; }
         public required string first_name { get; set; }
         public required string last_name { get; set; }
-        public required float balance { get; set; }
+        public required decimal balance { get; set; }
         public required string ticket_id { get; set; }
     }
 }
